+++ packages/splunk_otel_flutter/splunk_otel_flutter_platform_interface/lib/src/platform_interface/splunk_otel_flutter_platform_interface.dart
@@ -141,6 +141,14 @@ abstract class SplunkOtelFlutterPlatformInterface extends PlatformInterface {
   });
   Future<int> customTrackingStartWorkflow({required String workflowName});
   Future<void> customTrackingEndWorkflow({required int handle});
+  Future<void> customTrackingTrackError({
+    required String type,
+    required String message,
+    String? stacktrace,
+    required MutableAttributes attributes,
+    required String source,
+    required bool handled,
+  });
 
   // Navigation